Started testing Biggzlar's Breathe.cs & TPC
*Still using Third Person Camera asset script, but it's now following an
object that tracks player but uses the Breathe script from Biggzlar's
code to smooth out the tracking so it's not as jolty.
*Biggzlar's code doesn't use Rigidbody physics, which seems to make it
significantly less erratic applied to the duck character. It should be
less complicated to get my character to work properly if realistic-ish
physics aren't part of it (easier to fudge things into place), so I'm
liking that direction
*However, right now my duck is using Biggzlar's animations because I had
to use his animation controller; the Unity one uses blend trees and is a
bit more sophisticated, and I think I can borrow from that to make
animations work nicely
*Another issue is that walking on stairsteps makes the mesh abruptly
shift location, should at least tween the transition or something