Added FadeIn/Out, KB Prompt, Interior, and more
-Fader canvas element with an Animator Controller; triggers cause the
element to fade in or out
-Made 'Z' key act as a Submit button in Input settings so that Fungus
would accept it as a means of closing narrative captions just like A
button, Enter and Space
-A slow coroutine loop keeps tabs on whether a gamepad is connected in
HotspotManager.cs; if it is, gamepad prompts (currently just for A
button) will be shown, otherwise keyboard prompt for Z key will be
shown. It doesn't check last used input device, it simply will always
show gamepad prompts while a gamepad is connected.
-Teleporter.cs allows the character and camera target to be moved to a
new location for changing 'scenes'; I added an example apartment
interior for demonstration