NearestGround implemented, less janky teleporting
-Made a test scene using Gizmos and implemented NearestGround utility -
a vanilla class with a public method that uses a raycast to determine
the nearest "ground" (collider hit point in Vector3.down direction) to a
given Vector3 position.
-I made it primarily so that Teleporter.cs could warp player to the
nearest solid footing to the desired transform, so that Duckyboy would
be standing on the ground after teleporting instead of "falling down" to
it. It seems to work basically as it should, but there's still the
tiniest hint of jerkyness upon teleporting that probably has to do with
how his feet rest (or don't quite) on the transform's pivot point... but
I'm gonna call it a day on this issue for now, because it seems to have
less to do with my code and more to do with Duckyboy's model. (I guess I
could test that with another player model like Ethan...)
-It was fun using Gizmos to test the calculation, and it would be nice
to make an Editor script where I can 1) "Drop To Ground" - drop a
GameObject to the nearest ground on command (i.e. Ctrl+G or something),
or 2) "Stick To Ground" - have it stick persistently to the nearest
ground while translating (which would probably use a certain distance
above the GameObject to make the calculation, since the player
gameObject might otherwise be temporarily located below the nearest
ground at any position) - that would be really useful for positioning
things like scenery by dragging an object across the floor instead of
manually moving one axis at a time